Some cleanup helper functions
Helper functions for when I screw things up.

+++ server/proposal/cleanup.py
@@ -3,7 +3,7 @@ import os
 
 from django.conf import settings
 
-from .models import Image
+from .models import Attribute, Image
 
 logger = logging.getLogger(__name__)
 
@@ -14,8 +14,24 @@ def cleanup(rootpath=settings.DOC_ROOT):
 
         if os.path.exists(images_path):
             for file_name in os.listdir(images_path):
+                if file_name.startswith("thumb"):
+                    continue
+
                 file_path = os.path.join(images_path, file_name)
 
                 if not Image.objects.filter(image=file_path).exists():
                     logger.info("Removing unused image file: %s", file_path)
                     os.unlink(file_path)
+
+def remove_duplicate_attributes():
+    ids = set()
+    for a in Attribute.objects.all():
+        if a.pk in ids:
+            continue
+
+        attrs = Attribute.objects.filter(name=a.name,
+                                         proposal=a.proposal)\
+                                 .exclude(pk=a.pk)
+        ids.update([attr.pk for attr in attrs])
+
+    Attribute.objects.filter(pk__in=ids).delete()
